refactor : 딜러와 플레이어들을 Participants로 변경
