+++ src/main/java/blackJack/domain/participant/Player.java
@@ -4,6 +4,7 @@ public class Player extends Participant {
 
     private static final String DEALER_NAME = "딜러";
     private static final String ERROR_MESSAGE_PROHIBIT_NAME = "플레이어의 이름은 '딜러'일 수 없습니다.";
+    private static final int BLACK_JACK = 21;
 
     public Player(String name) {
         super(name);
@@ -14,5 +15,10 @@ public class Player extends Participant {
         if (name.equals(DEALER_NAME)) {
             throw new IllegalArgumentException(ERROR_MESSAGE_PROHIBIT_NAME);
         }
+    }
+
+    @Override
+    public boolean hasNextTurn() {
+        return this.getScore() <= BLACK_JACK;
     }
 }
