+++ src/test/java/blackJack/domain/BlackJackGameTest.java
@@ -7,40 +7,17 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import blackJack.domain.card.Card;
-import blackJack.domain.card.Denomination;
-import blackJack.domain.card.Symbol;
 import blackJack.domain.participant.Dealer;
+import blackJack.domain.participant.Participants;
 import blackJack.domain.participant.Player;
 
 class BlackJackGameTest {
 
     @Test
-    @DisplayName("Participants 생성 테스트")
+    @DisplayName("BlackJackGame 생성 테스트")
     void createValidDealer() {
-        assertThat(new BlackJackGame(new Dealer(), List.of(new Player("rookie")))).isNotNull();
-    }
-
-    @Test
-    @DisplayName("Player들의 이름이 중복된 경우 테스트")
-    void checkDuplicatePlayerName() {
-        List<Player> players = List.of(new Player("rookie"), new Player("rookie"));
-
-        assertThatThrownBy(() -> new BlackJackGame(new Dealer(), players))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("플레이어의 이름은 중복될 수 없습니다.");
-    }
-
-    @Test
-    @DisplayName("Player의 수가 1-7명이 아닌 경우 테스트")
-    void checkPlayerCount() {
-        List<Player> players = List.of(
-            new Player("k1"), new Player("k2"), new Player("k3"), new Player("k4"),
-            new Player("k5"), new Player("k6"), new Player("k7"), new Player("k8"));
-
-        assertThatThrownBy(() -> new BlackJackGame(new Dealer(), players))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("플레이어의 인원수는 1명 이상 7명 이하여야 합니다.");
+        Participants participants = new Participants(new Dealer(), List.of(new Player("rookie")));
+        assertThat(new BlackJackGame(participants)).isNotNull();
     }
 
     @Test
@@ -48,61 +25,11 @@ class BlackJackGameTest {
     void firstCardDispensing() {
         Player player1 = new Player("kei");
         Player player2 = new Player("rookie");
+        Participants participants = new Participants(new Dealer(), List.of(player1, player2));
 
-        BlackJackGame blackJackGame = new BlackJackGame(new Dealer(), List.of(player1, player2));
+        BlackJackGame blackJackGame = new BlackJackGame(participants);
         blackJackGame.firstCardDispensing();
 
         assertThat(player1.getCards().size()).isEqualTo(2);
-    }
-
-    @Test
-    @DisplayName("플레이어의 카드 추가 분배가 불가능한 경우 테스트")
-    void hasFalsePlayerNextTurn() {
-        Player player = new Player("kei");
-        player.receiveCard(new Card(Symbol.SPADE, Denomination.J));
-        player.receiveCard(new Card(Symbol.HEART, Denomination.J));
-        player.receiveCard(new Card(Symbol.SPADE, Denomination.TWO));
-
-        BlackJackGame blackJackGame = new BlackJackGame(new Dealer(), List.of(player));
-
-        assertThat(blackJackGame.hasNextTurn(player)).isFalse();
-    }
-
-    @Test
-    @DisplayName("플레이어의 카드 추가 분배가 가능한 경우 테스트")
-    void hasTruePlayerNextTurn() {
-        Player player = new Player("kei");
-        player.receiveCard(new Card(Symbol.SPADE, Denomination.J));
-        player.receiveCard(new Card(Symbol.HEART, Denomination.J));
-
-        BlackJackGame blackJackGame = new BlackJackGame(new Dealer(), List.of(player));
-
-        assertThat(blackJackGame.hasNextTurn(player)).isTrue();
-    }
-
-    @Test
-    @DisplayName("딜러의 카드 추가 분배가 불가능한 경우 테스트")
-    void hasFalseDealerNextTurn() {
-        Player player = new Player("kei");
-        Dealer dealer = new Dealer();
-        dealer.receiveCard(new Card(Symbol.SPADE, Denomination.SEVEN));
-        dealer.receiveCard(new Card(Symbol.HEART, Denomination.J));
-
-        BlackJackGame blackJackGame = new BlackJackGame(new Dealer(), List.of(player));
-
-        assertThat(blackJackGame.hasNextTurn(dealer)).isFalse();
-    }
-
-    @Test
-    @DisplayName("딜러의 카드 추가 분배가 가능한 경우 테스트")
-    void hasTrueDealerNextTurn() {
-        Player player = new Player("kei");
-        Dealer dealer = new Dealer();
-        dealer.receiveCard(new Card(Symbol.SPADE, Denomination.SIX));
-        dealer.receiveCard(new Card(Symbol.HEART, Denomination.J));
-
-        BlackJackGame blackJackGame = new BlackJackGame(new Dealer(), List.of(player));
-
-        assertThat(blackJackGame.hasNextTurn(dealer)).isTrue();
     }
 }
